Added audio to notify timer end. Changed timer to show sessions over 60 minutes.
Audio file was added to HTML and will be played once the time on timer reaches 0.  The audio is stopped when user clicks anywhere on the screen.

Timer used to only display minutes and seconds. Added time conversion for day and hour. Timer now will display the hour if the user chooses a session time over 60 minutes.